chore: Remove hardcoded insights
